Management Procedures

Management Procedures

Usage

Create_Rec_List(DataList)

Get_Year_Info(DataList)

Adjust_Effort(
  RecList,
  DataList,
  Effort_Mod = 0,
  First_Management_Year = 2025,
  ...
)

Close_Areas(
  RecList,
  DataList,
  areas = c(1, 1, 1, 1, 1, 1),
  Allocate = 1,
  First_Management_Year
)

Effort_Creep(
  RecList,
  DataList,
  First_Management_Year = 2025,
  effort_creep = 0.02,
  ...
)

SQ(x, DataList, ...)

SQ_NS(
  x,
  DataList,
  First_Management_Year = 2025,
  areas = c(1, 0, 1, 0, 1, 0),
  Allocate = 1,
  ...
)

SQ_OS(
  x,
  DataList,
  First_Management_Year = 2025,
  areas = c(0, 1, 0, 1, 0, 1),
  Allocate = 1,
  ...
)

SQ_FR(x, DataList, First_Management_Year = 2025, ...)

SQ_FR_NS(
  x,
  DataList,
  First_Management_Year = 2025,
  areas = c(1, 0, 1, 0, 1, 0),
  Allocate = 1,
  ...
)

SQ_FR_OS(
  x,
  DataList,
  First_Management_Year = 2025,
  areas = c(0, 1, 0, 1, 0, 1),
  Allocate = 1,
  ...
)

SQ_MLL(x, DataList, First_Management_Year = 2025, ...)

SQ_MLL_NS(
  x,
  DataList,
  First_Management_Year = 2025,
  areas = c(1, 0, 1, 0, 1, 0),
  Allocate = 1,
  ...
)

SQ_MLL_OS(
  x,
  DataList,
  First_Management_Year = 2025,
  areas = c(0, 1, 0, 1, 0, 1),
  Allocate = 1,
  ...
)

SQ_FR_MLL(x, DataList, First_Management_Year = 2025, ...)

SQ_FR_MLL_NS(
  x,
  DataList,
  First_Management_Year = 2025,
  areas = c(1, 0, 1, 0, 1, 0),
  Allocate = 1,
  ...
)

SQ_FR_MLL_OS(
  x,
  DataList,
  First_Management_Year = 2025,
  areas = c(0, 1, 0, 1, 0, 1),
  Allocate = 1,
  ...
)

Arguments

DataList

A nested list of Data objects

RecList

A list of Rec objects returned by an MMP

Effort_Mod

Effort modifier. Effort is adjusted as Effort_Rec_Fleet <- Effort_Rec_Fleet * (1-Effort_Mod)

First_Management_Year

Year the management procedures are first implemented

...

Additional arguments

x

Simulation number

Functions

Create_Rec_List(): Generate a list of Rec objects for an MMP

Get_Year_Info(): Return information on current year from a DataList

Adjust_Effort(): Adjust fishing effort for General Recreational Fleet

Close_Areas(): Close areas to fishing

Effort_Creep(): 2% annual creep in fishing effort for Gen Rec fleet

SQ(): Status Quo. Fishing effort for each fleet is fixed in the projection years at the geometric mean of the three last historical years

SQ_NS(): Status Quo with offshore areas closed to fishing

SQ_OS(): Status Quo with nearshore areas closed to fishing

SQ_FR(): Status Quo with Full Retention. Modify retention curve with an asymptote of 1

SQ_FR_NS(): Status Quo with full retention and offshore closed

SQ_FR_OS(): Status Quo with full retention and nearshore closed

SQ_MLL(): Status Quo with a stock specific size limit.

SQ_MLL_NS(): Status Quo with a stock specific size limit and offshore areas closed to fishing

SQ_MLL_OS(): Status Quo with a stock specific size limit and nearshore areas closed to fishing

SQ_FR_MLL(): Status Quo with full retention and minimum size limit

SQ_FR_MLL_NS(): Status Quo with full retention, minimum size limit, and offshore closed

SQ_FR_MLL_OS(): Status Quo with full retention, minimum size limit, and nearshore closed